+++ stalkerrender/source/base/FVisual.cpp
@@ -128,14 +128,14 @@ void Fvisual::Load		(const char* N, IReader *data, u32 dwFlags)
 
 #if defined(USE_DX10) || defined(USE_DX11)
 			VERIFY				(NULL==p_rm_Vertices);
-			R_CHK				(dx10BufferUtils::CreateVertexBuffer(&p_rm_Vertices, data->pointer(), vCount*vStride));
+			R_CHK				(dx10BufferUtils::CreateVertexBuffer(&p_rm_Vertices, data->pointer(), static_cast<UINT>(vCount*vStride)));
 			HW.stats_manager.increment_stats_vb						(p_rm_Vertices);
 #else	//	USE_DX10
 			BOOL	bSoft		= HW.Caps.geometry.bSoftware;
 			u32		dwUsage		= D3DUSAGE_WRITEONLY | (bSoft?D3DUSAGE_SOFTWAREPROCESSING:0);
 			BYTE*	bytes		= 0;
 			VERIFY				(NULL==p_rm_Vertices);
-			R_CHK				(HW.pDevice->CreateVertexBuffer	(vCount*vStride,dwUsage,0,D3DPOOL_MANAGED,&p_rm_Vertices,0));
+			R_CHK				(HW.pDevice->CreateVertexBuffer	(static_cast<UINT>(vCount*vStride),dwUsage,0,D3DPOOL_MANAGED,&p_rm_Vertices,0));
 			HW.stats_manager.increment_stats_vb					(p_rm_Vertices);
 			R_CHK				(p_rm_Vertices->Lock(0,0,(void**)&bytes,0));
 			CopyMemory			(bytes, data->pointer(), vCount*vStride);
@@ -175,7 +175,7 @@ void Fvisual::Load		(const char* N, IReader *data, u32 dwFlags)
 			//CopyMemory		(bytes, data->pointer(), iCount*2);
 
 			VERIFY				(NULL==p_rm_Indices);
-			R_CHK				(dx10BufferUtils::CreateIndexBuffer(&p_rm_Indices, data->pointer(), iCount*2));
+			R_CHK				(dx10BufferUtils::CreateIndexBuffer(&p_rm_Indices, data->pointer(), static_cast<UINT>(iCount*2)));
 			HW.stats_manager.increment_stats_ib		( p_rm_Indices);
 #else	//	USE_DX10
 			BOOL	bSoft		= HW.Caps.geometry.bSoftware;
@@ -183,7 +183,7 @@ void Fvisual::Load		(const char* N, IReader *data, u32 dwFlags)
 			BYTE*	bytes		= 0;
 
 			VERIFY				(NULL==p_rm_Indices);
-			R_CHK				(HW.pDevice->CreateIndexBuffer(iCount*2,dwUsage,D3DFMT_INDEX16,D3DPOOL_MANAGED,&p_rm_Indices,0));
+			R_CHK				(HW.pDevice->CreateIndexBuffer(static_cast<UINT>(iCount*2),dwUsage,D3DFMT_INDEX16,D3DPOOL_MANAGED,&p_rm_Indices,0));
 			HW.stats_manager.increment_stats_ib		( p_rm_Indices);
 			R_CHK				(p_rm_Indices->Lock(0,0,(void**)&bytes,0));
 			CopyMemory		(bytes, data->pointer(), iCount*2);
