+++ stalkerrender/source/r4/CSCompiler.cpp
@@ -194,7 +194,7 @@ void CSCompiler::compile(const char* name)
 	LPCSTR						c_target	= "cs_5_0";
 	LPCSTR						c_entry		= "main";
 
-	HRESULT	const _hr			= ::Render->shader_compile(name,(DWORD const*)file->pointer(),file->length(), c_entry, c_target, D3D10_SHADER_PACK_MATRIX_ROW_MAJOR, (void*&)m_cs );
+	HRESULT	const _hr			= ::Render->shader_compile(name,(DWORD const*)file->pointer(), static_cast<UINT>(file->length()), c_entry, c_target, D3D10_SHADER_PACK_MATRIX_ROW_MAJOR, (void*&)m_cs );
 
 	XRayBearReader::Destroy(file);
